[OGUI-1239] Refactor utility file and include tests (#1502)
* Update default eslint and config files
* Update package file to ignore lint for one of the jobs
* Update utility file and tests associated to it
* Update GH workflow
