+++ QualityControl/config-default.js
@@ -10,28 +10,29 @@
  * In applying this license CERN does not waive the privileges and immunities
  * granted to it by virtue of its status as an Intergovernmental Organization
  * or submit itself to any jurisdiction.
-*/
+ */
 
-export default {
-  // demoData: false,
+export const config = {
+  demoData: false,
 
   http: {
     port: 8080,
-    // portSecure: 8443,
+    // PortSecure: 8443,
     hostname: 'localhost',
-    // key: './cert/key.pem',
-    // cert: './cert/cert.pem',
-    tls: false
+
+    /*
+     * Key: './cert/key.pem',
+     * cert: './cert/cert.pem',
+     */
+    tls: false,
   },
 
   mysql: {
     host: 'localhost',
     user: 'qc_user',
     password: 'qc_user',
-    database: 'quality_control'
+    database: 'quality_control',
   },
-
-  listingConnector: 'ccdb', // ccdb or mysql or amore (default mysql) for listing objects
 
   ccdb: {
     protocol: 'http',
@@ -40,48 +41,52 @@ export default {
     prefix: 'qc',
   },
 
-  // consul: {
-  //   hostname: 'localhost',
-  //   port: 8500,
-  //   refreshRate: {
-  //     min: 10,
-  //     max: 120
-  //   }
-  // }
+  /*
+   * Consul configuration object
+   * consul: {
+   *   hostname: 'localhost',
+   *   port: 8500,
+   *   refreshRate: {
+   *     min: 10,
+   *     max: 120
+   *   }
+   * }
+   */
 
-  // dbFile: '/var/db/qcg-db.json', // absolute path where to save layouts, default = root of this app
+  /*
+   * Absolute path where to save layouts, default = root of this app
+   * dbFile: '/var/db/qcg-db.json'
+   */
 
-  // JWT manages user's session duration
-  // https://github.com/AliceO2Group/WebUi/blob/dev/Framework/docs/guide/json-tokens.md
-  // jwt: {
-  //   secret: '<secret>',
-  //   issuer: 'alice-o2-gui',
-  //   expiration: '1d',
-  //   maxAge: '1d'
-  // },
+  /*
+   * JWT manages user's session duration
+   * https://github.com/AliceO2Group/WebUi/blob/dev/Framework/docs/guide/json-tokens.md
+   * jwt: {
+   *   secret: '<secret>',
+   *   issuer: 'alice-o2-gui',
+   *   expiration: '1d',
+   *   maxAge: '1d'
+   * },
+   */
 
-  // oAuth: {
-  //   secret: '<oauth secret>',
-  //   id: '<oauth name>',
-  //   tokenHost: 'https://oauth.web.cern.ch',
-  //   tokenPath: '/OAuth/Token',
-  //   authorizePath: '/OAuth/Authorize',
-  //   redirect_uri: 'https://<Gui url>/callback',
-  //   scope: 'https://oauthresource.web.cern.ch/api/User',
-  //   state: '3(#0/!~',
-  //   egroup: 'alice-member',
-  //   resource: {
-  //     hostname: 'oauthresource.web.cern.ch',
-  //     userPath: '/api/User',
-  //     groupPath: '/api/Groups',
-  //     port: 443
-  //   }
-  // },
-
-  // amore: {
-  //   host: '<mysql host>',
-  //   user: '<login>',
-  //   password: '<pwd>',
-  //   database: 'AMORE'
-  // }
+  /*
+   * CERN OpenID configuration object
+   * openId: {
+   *   secret: '<oauth secret>',
+   *   id: '<oauth name>',
+   *   tokenHost: 'https://oauth.web.cern.ch',
+   *   tokenPath: '/OAuth/Token',
+   *   authorizePath: '/OAuth/Authorize',
+   *   redirect_uri: 'https://<Gui url>/callback',
+   *   scope: 'https://oauthresource.web.cern.ch/api/User',
+   *   state: '3(#0/!~',
+   *   egroup: 'alice-member',
+   *   resource: {
+   *     hostname: 'oauthresource.web.cern.ch',
+   *     userPath: '/api/User',
+   *     groupPath: '/api/Groups',
+   *     port: 443
+   *   }
+   * },
+   */
 };
