+++ QualityControl/test/mocha-index.js
@@ -11,12 +11,11 @@
  * granted to it by virtue of its status as an Intergovernmental Organization
  * or submit itself to any jurisdiction.
  */
+'use strict';
+import utilsTestSuite from './lib/utils/utils.test.js';
+import configurationTestSuite from './lib/config/public-config.test.js';
 
 describe('Lib Test Suite', async () => {
-  try {
-    require('./public');
-    // Require('./lib');
-  } catch (error) {
-    console.error(error);
-  }
+  describe('Utility methods test suite', async () => await utilsTestSuite());
+  describe('Configuration File Parser test suite', async () => await configurationTestSuite());
 });
